[bugfix] call parent method in eglot-initialization-options
(eglot-initialization-options) is a cl-defgeneric overridable method with a default implementation, and our code was overriding it but not calling the default implementation. 

This means that any core eglot logic around initialization options was failing to run, which had a user-visible consequence that the initializationOptions behaviour described in https://joaotavora.github.io/eglot/#User_002dspecific-configuration was not happening.